requirements.txt
Django==1.10.5
appdirs==1.4.2
django-betterforms==1.1.4
django-formtools==2.0
django-localflavor==1.4.1
mysqlclient==1.3.10
packaging==16.8
pyparsing==2.1.10
pytz==2016.10
six==1.10.0